Added the file `working report.Rmd`, and completed Part 0
Added the file `working report.Rmd`, which is the working report for this project. This version contains the code for Part 0, in which I determined how R reads/differentiates between Hiragana and Katakana through a short experiment. 

Continuations of this project should be done in subsequent parts under the Part 0 section.